Move celery tasks to package
